替换timer为quartz，使用jetCache缓存数据，修复bug

+++ src/main/java/com/meeting/intelligent/service/MeetingRoomTypeService.java
@@ -16,7 +16,7 @@ public interface MeetingRoomTypeService extends IService<MeetingRoomTypeEntity> 
 
     PageUtils queryPage(Map<String, Object> params);
 
-    void checkTypeExist(Integer typeId);
+    void checkTypeExist(Long typeId);
 
     void deleteTypes(List<Long> typeIds);
 }
